requirements.txt
Django>=5.0
djangorestframework>=3.14.0
botocore>=1.34.10